+++ yk_face/face.py
@@ -5,6 +5,7 @@ from typing import List, Dict
 from yk_utils.images import parse_image
 from yk_utils.apis import request, request_async
 from yk_face_api_models import ProcessRequest, VerifyRequest, VerifyIdRequest, IdentifyRequest
+from yk_face.util import FaceException, face_process_validation
 
 
 @dataclass
@@ -206,3 +207,28 @@ async def identify_async(
         gallery_id=group_id
     ).to_dict()
     return await request_async('POST', FaceRouterEndpoints.identify, json=identify_request)
+
+
+def verify_images(first_image, second_image) -> float:
+    """
+        Verifies if the face detected on the first image matches to the
+        face detected on the second image.
+    :param first_image:
+        A base64 string or a file path or a file-like object representing an image.
+    :param second_image:
+        A base64 string or a file path or a file-like object representing an image.
+    :return:
+        Matching Score.
+    """
+    first_face = process(first_image)
+    second_face = process(second_image)
+
+    error_message = face_process_validation(first_face)
+    if error_message:
+        raise FaceException(f"First image: {error_message}")
+
+    error_message = face_process_validation(second_face)
+    if error_message:
+        raise FaceException(f"Second image: {error_message}")
+
+    return verify(first_face[0]["template"], second_face[0]["template"])